fixed bug in bin.* dealing with returning number in bin as int and not double. HoD works for XY2D trajectory.

+++ bin.cpp
@@ -33,7 +33,7 @@ double Bin::getFreq()
 	return freq;
 }
 
-int Bin::getBinSize()
+double Bin::getBinSize()
 {
 	return numInBin;
 }
